chore: Refactor login and signup components
